+++ client/src/tests/mod.rs
@@ -7,6 +7,7 @@ use opcua_types::MessageSecurityMode;
 use opcua_core::crypto::SecurityPolicy;
 
 use config::{ClientConfig, ClientEndpoint, ClientUserToken, ANONYMOUS_USER_TOKEN_ID};
+use client::ClientBuilder;
 
 fn make_test_file(filename: &str) -> PathBuf {
     let mut path = std::env::temp_dir();
@@ -15,55 +16,44 @@ fn make_test_file(filename: &str) -> PathBuf {
 }
 
 pub fn default_sample_config() -> ClientConfig {
-    use std::path::PathBuf;
     use opcua_core::crypto::SecurityPolicy;
     use opcua_types::MessageSecurityMode;
 
-    let pki_dir = PathBuf::from("./pki");
-    let mut endpoints = BTreeMap::new();
-    endpoints.insert(String::from("sample_none"), ClientEndpoint {
-        url: String::from("opc.tcp://127.0.0.1:4855/"),
-        security_policy: String::from(SecurityPolicy::None.to_str()),
-        security_mode: String::from(MessageSecurityMode::None),
-        user_token_id: ANONYMOUS_USER_TOKEN_ID.to_string(),
-    });
-    endpoints.insert(String::from("sample_basic128rsa15"), ClientEndpoint {
-        url: String::from("opc.tcp://127.0.0.1:4855/"),
-        security_policy: String::from(SecurityPolicy::Basic128Rsa15.to_str()),
-        security_mode: String::from(MessageSecurityMode::SignAndEncrypt),
-        user_token_id: ANONYMOUS_USER_TOKEN_ID.to_string(),
-    });
-    endpoints.insert(String::from("sample_basic256"), ClientEndpoint {
-        url: String::from("opc.tcp://127.0.0.1:4855/"),
-        security_policy: String::from(SecurityPolicy::Basic256.to_str()),
-        security_mode: String::from(MessageSecurityMode::SignAndEncrypt),
-        user_token_id: ANONYMOUS_USER_TOKEN_ID.to_string(),
-    });
-    endpoints.insert(String::from("sample_basic256sha256"), ClientEndpoint {
-        url: String::from("opc.tcp://127.0.0.1:4855/"),
-        security_policy: String::from(SecurityPolicy::Basic256Sha256.to_str()),
-        security_mode: String::from(MessageSecurityMode::SignAndEncrypt),
-        user_token_id: ANONYMOUS_USER_TOKEN_ID.to_string(),
-    });
-    let mut user_tokens = BTreeMap::new();
-    user_tokens.insert(
-        String::from("sample_user"),
-        ClientUserToken {
-            user: String::from("sample"),
-            password: String::from("sample1")
-        });
-    ClientConfig {
-        application_name: "OPC UA Sample Client".to_string(),
-        application_uri: "urn:SampleClient".to_string(),
-        create_sample_keypair: true,
-        trust_server_certs: true,
-        product_uri: String::new(),
-        pki_dir,
-        preferred_locales: Vec::new(),
-        default_endpoint: "sample_none".to_string(),
-        endpoints,
-        user_tokens
-    }
+    ClientBuilder::new()
+        .application_name("OPC UA Sample Client")
+        .application_uri("urn:SampleClient")
+        .pki_dir("./pki")
+        .endpoints(vec![
+            ("sample_none", ClientEndpoint {
+                url: String::from("opc.tcp://127.0.0.1:4855/"),
+                security_policy: String::from(SecurityPolicy::None.to_str()),
+                security_mode: String::from(MessageSecurityMode::None),
+                user_token_id: ANONYMOUS_USER_TOKEN_ID.to_string(),
+            }),
+            ("sample_basic128rsa15", ClientEndpoint {
+                url: String::from("opc.tcp://127.0.0.1:4855/"),
+                security_policy: String::from(SecurityPolicy::Basic128Rsa15.to_str()),
+                security_mode: String::from(MessageSecurityMode::SignAndEncrypt),
+                user_token_id: ANONYMOUS_USER_TOKEN_ID.to_string(),
+            }),
+            ("sample_basic256", ClientEndpoint {
+                url: String::from("opc.tcp://127.0.0.1:4855/"),
+                security_policy: String::from(SecurityPolicy::Basic256.to_str()),
+                security_mode: String::from(MessageSecurityMode::SignAndEncrypt),
+                user_token_id: ANONYMOUS_USER_TOKEN_ID.to_string(),
+            }),
+            ("sample_basic256sha256", ClientEndpoint {
+                url: String::from("opc.tcp://127.0.0.1:4855/"),
+                security_policy: String::from(SecurityPolicy::Basic256Sha256.to_str()),
+                security_mode: String::from(MessageSecurityMode::SignAndEncrypt),
+                user_token_id: ANONYMOUS_USER_TOKEN_ID.to_string(),
+            })
+        ])
+        .default_endpoint("sample_none")
+        .create_sample_keypair(true)
+        .trust_server_certs(true)
+        .user_token("sample_user", ClientUserToken::new("sample", "sample1"))
+        .config()
 }
 
 #[test]
@@ -140,7 +130,7 @@ fn client_anonymous_user_tokens_id() {
         String::from("ANONYMOUS"),
         ClientUserToken {
             user: String::new(),
-            password: String::new()
+            password: String::new(),
         });
     assert!(!config.is_valid());
 }